refactor: 인기 테마 조회 기능을 Theme 도메인으로 이동하고 관련 계층을 재구성
