+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,21 +1,15 @@
 package roomescape.controller;
 
 import java.net.URI;
-import java.time.LocalDate;
-import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.ReservationRequestDTO;
 import roomescape.dto.ReservationResponseDTO;
-import roomescape.dto.ReservationTimeResponseDTO;
-import roomescape.dto.ThemeResponseDTO;
 import roomescape.service.ReservationService;
 
 @RestController
@@ -26,12 +20,6 @@ public class ReservationController {
 
     public ReservationController(ReservationService reservationService) {
         this.reservationService = reservationService;
-    }
-
-    @GetMapping("/popular-themes")
-    public ResponseEntity<List<ThemeResponseDTO>> getPopularThemes() {
-        List<ThemeResponseDTO> popularThemes = reservationService.getPopularThemes();
-        return ResponseEntity.ok(popularThemes);
     }
 
     @PostMapping
