+++ src/main/java/roomescape/repository/ThemeRepository.java
@@ -12,5 +12,7 @@ public interface ThemeRepository {
 
     Optional<Theme> findById(Long id);
 
+    List<Long> findPopularThemeIds();
+
     void delete(Long id);
 }
