+++ src/test/java/roomescape/RoomescapeApplicationTest.java
@@ -109,9 +109,10 @@ class RoomescapeApplicationTest {
     @Sql("/data.sql")
     void 인기_테마_조회() {
         RestAssured.given().log().all()
-                .when().get("/reservations/popular-themes")
+                .when().get("/themes/popular-themes")
                 .then().log().all()
                 .statusCode(200)
                 .body("id", contains(1, 2, 3, 6, 5, 4, 8, 7, 10, 9));
     }
+
 }
